backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlmodel>=0.0.22
argon2-cffi>=23.1
itsdangerous>=2.2
pydantic-settings>=2.6
email-validator>=2.2
python-multipart>=0.0.12
